input state corrected with checked status
