+++ components/toolbar/components/filters/components/filterList.js
@@ -1,0 +1,56 @@
+const filters = [
+  {
+    "_worksheetName": "Profit Ratio Map",
+    "_fieldName": "Segment",
+    "_filterType": "categorical",
+    "_fieldId": "[federated.1d5vk860twkhqd1bvit1i02ev3x0].[none:Country:nk]",
+    "_registryId": 0,
+    "_appliedValues": [
+      {
+        "_value": "Consumer",
+        "_nativeValue": "Consumer",
+        "_formattedValue": "Consumer"
+      },
+      {
+        "_value": "Corporate",
+        "_nativeValue": "Corporate",
+        "_formattedValue": "Corporate"
+      },
+      {
+        "_value": "Home Office",
+        "_nativeValue": "Home Office",
+        "_formattedValue": "Home Office"
+      },
+    ],
+    "_isExcludeMode": false,
+    "_isAllSelected": true
+  },
+  {
+    "_worksheetName": "Profit Ratio Map",
+    "_fieldName": "Product Category",
+    "_filterType": "categorical",
+    "_fieldId": "[federated.1d5vk860twkhqd1bvit1i02ev3x0].[none:Country:nk]",
+    "_registryId": 0,
+    "_appliedValues": [
+      {
+        "_value": "Technology",
+        "_nativeValue": "Technology",
+        "_formattedValue": "Technology"
+      },
+      {
+        "_value": "Office Supplies",
+        "_nativeValue": "Office Supplies",
+        "_formattedValue": "Office Supplies"
+      },
+      {
+        "_value": "Furniture",
+        "_nativeValue": "Furniture",
+        "_formattedValue": "Furniture"
+      },
+    ],
+    "_isExcludeMode": false,
+    "_isAllSelected": true
+  }
+]
+
+export default filters;
